Changed a few irregularities and also added markers. Also fixed some examples.
First issue fixed was that the CHANGE command could not use sqrt, but
it could use other math operators. Also, NSEQU (not equal strings) was
not working at all.

Finally, the most important thing I added was markers. Markers are a
way to save points in the code and then go back to them later in the
code. This can help with making loops, by using markers with if
statements. For an example, please see examples/wait_for_password.bit.